Create palindromeString
This program will check if the given input string is palindrome or not.

A string is said to be a palindrome if the string read from left to right is equal to the string read from right to left. 
For example, ignoring the difference between uppercase and lowercase letters, the string "madam" is a palindrome, while the string "madAm" is not so. 

The following is an inductive characterization of a palindrome string: 
•the empty string is a palindrome; 
•a string constituted only by a single character is a palindrome; 
•a string c s d is a palindrome, if s is a palindrome and c is a character equal to d; 
•nothing else is a palindrome.